BackEnd/Domains/Users/SearchUser/requirements.txt
absl-py==2.1.0
aniso8601==7.0.0
astunparse==1.6.3
attrs==25.1.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
flasgger==0.9.7.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-GraphQL==2.0.1
flask-swagger-ui==4.11.1
flatbuffers==24.12.23
fonttools==4.55.3
gast==0.6.0
google-pasta==0.2.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
grpcio==1.69.0
h5py==3.12.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mistune==3.1.1
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pillow==11.1.0
promise==2.3
protobuf==5.29.3
psycopg2-binary==2.9.10
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
Rx==1.6.3
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
